fix(objectstore): share signed-request handling across providers
